// src/schemas/academicManagement.schema.ts
import { z } from "zod";

export const academicSemesterSchema = z.object({
  name: z.string({ required_error: "Please selete a name" }),
  year: z.string({ required_error: "Please selete a year" }),
  startMonth: z.string({ required_error: "Please selete a start month" }),
  endMonth: z.string({ required_error: "Please selete a end month" }),
});